+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Each individual change should have a link to the pull request after the descript
 Changed
 ^^^^^^^
 - restricted pandas versions to < 3
+- converted ColumnDtypeSetter to narwhals/lazyframe/json
 
 2.7.0 (20/01/2026)
 ------------------